pyproject.toml
[project]
name = "stripe-copy-prod-to-test"
version = "1.0.0"
description = "Script pour copier les produits et tarifs Stripe de production vers test"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "stripe>=8.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]

[project.scripts]
stripe-copy = "main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py38"
